Extract log-based replication test into a different file and convert it to jest
Extract log-based replication test into a different file and convert it to jest
